+++ Models/Productos.cs
@@ -2,5 +2,5 @@ public class Productos
 {
     public int idProducto { get; set; }
     public string Descripcion {get; set;}
-    public int Precio { get; set; }
+    public Decimal Precio { get; set; }
 }
